Update README.md
Beyond Compare Crack  Is a reliable and powerful program that lets you compare different file formats quickly and easily, merge documents into one, and keep folders in sync. It is a full application that comes with a lot of different tools for comparing files that let you do a thorough byte-to-byte comparison.